Move FormulaEngine and formula generators into a separate package
