+++ src/frequenz/sdk/timeseries/_formula_engine/_formula_generators/_battery_soc_formula.py
