+++ src/frequenz/sdk/timeseries/_formula_engine/_formula_generators/_grid_power_formula.py
