+++ src/frequenz/sdk/timeseries/_formula_engine/_formula_generators/_pv_power_formula.py
